requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
chromedriver-binary==95.0.4638.69.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.9
django-celery-results==2.2.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
kombu==5.2.1
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
prompt-toolkit==3.0.22
psycopg2==2.9.2
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
